+++ src/parser/ELR/builder/model/data.ts
@@ -12,4 +12,5 @@ export type ParserBuilderData<
 > = {
   defs: Definition<Kinds>;
   ctxBuilder?: DefinitionContextBuilder<ASTData, ErrorType, Kinds, LexerKinds>;
+  // TODO: add hydration id?
 };