+++ CHANGELOG.md
@@ -39,6 +39,7 @@
     - Add `checkRollback` for `IParserBuilder.build`.
   - **_Breaking Change_**: `IParserBuilder.entry` will check kinds.
   - **_Breaking Change_**: `IParserBuilder.build` will return `{ parser, serializable }`.
+  - **_Breaking Change_**: remove `IParserBuilder.leftSA/rightSA`, enhance the `IParserBuilder.priority` to support left-to-right or right-to-left associativity.
   - Feat: add `ASTNodeSelector`.
   - Feat: resolved conflicts in definition context will have higher priority than resolved conflicts defined in parser builder.
   - Feat: typed parser.